converted drawer navigation to stack navigation because of bug with babel plugin
