+++ navigation/RootStack.tsx
@@ -4,11 +4,7 @@ import { RootStackParamList } from "../types/types";
 import EntryScreen from '../screens/EntryScreen';
 import LogInScreen from '../screens/LogInScreen';
 import SignUpScreen from '../screens/SignUpScreen';
-import DashboardScreen from '../screens/DashboardScreen';
-import AllListsScreen from "../screens/AllListsScreen";
-import ListDetailScreen from "../screens/ListDetailScreen";
-import HeaderWithBackButton from "../components/HeaderWithBackButton";
-import HeaderNoBackButton from "../components/HeaderNoBackButton";
+import HomeStack from "./HomeStack";
 
 
 const Stack = createNativeStackNavigator<RootStackParamList>();
@@ -16,42 +12,25 @@ const Stack = createNativeStackNavigator<RootStackParamList>();
 export default function RootStack() {
 
   return (
-    <Stack.Navigator>
+    <Stack.Navigator 
+      initialRouteName="Entry"
+      screenOptions={{ headerShown: false }}
+    >
       <Stack.Screen
-        options={{ headerShown: false }}
         name="Entry"
         component={EntryScreen}
       />
       <Stack.Screen
-        options={{ headerShown: false }}
         name="Login"
         component={LogInScreen}
       />
       <Stack.Screen
-        options={{ headerShown: false }}
         name="SignUp"
         component={SignUpScreen}
       />
       <Stack.Screen
-        options={{
-          header: () => <HeaderNoBackButton />
-        }}
         name="Home"
-        component={DashboardScreen}
-      />
-      <Stack.Screen
-        options={{
-          header: () => <HeaderWithBackButton />,
-        }}
-        name="Lists"
-        component={AllListsScreen}
-      />
-      <Stack.Screen
-        options={{
-          header: () => <HeaderWithBackButton />
-        }}
-        name="ListDetail"
-        component={ListDetailScreen}
+        component={HomeStack}
       />
     </Stack.Navigator>
   )
